backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "python-jose[cryptography]>=3.3.0",
    "apscheduler>=3.11.0",
]
